requirements.txt
ultralytics>=8.2.0
streamlit>=1.35.0
opencv-python-headless>=4.9.0
torch>=2.2.0
pyyaml>=6.0
numpy>=1.26.0